Persist bowling game to file.
Added load() and save() to bowling_web.erl. These read and write the store dictionary from/to a file. Data is loaded on server start. Added 'stop' message to handler which saves the data and stops the server.
Added start_server.erl, an escript wrapper to set the code path, compile the bowling modules, and start the Spooky server.